[TIDY] (generator) Moved type related functions to their own file
