+++ .generator/src/generator/setup.py
@@ -4,6 +4,7 @@ import pathlib
 from . import openapi
 from . import formatter
 from . import utils
+from . import type
 
 
 def load_environment(version: str) -> Environment:
@@ -14,27 +15,27 @@ def load_environment(version: str) -> Environment:
     env.filters["attribute_name"] = formatter.attribute_name
     env.filters["camel_case"] = formatter.camel_case
     env.filters["sanitize_description"] = formatter.sanitize_description
-    env.filters["parameter_schema"] = openapi.parameter_schema
-    env.filters["parameters"] = openapi.parameters
-    env.filters["response_type"] = openapi.get_type_for_response
-    env.filters["return_type"] = openapi.return_type
     env.filters["simple_type"] = formatter.simple_type
     env.filters["snake_case"] = formatter.snake_case
     env.filters["untitle_case"] = formatter.untitle_case
-    env.filters["capitalize"] = utils.capitalize
     env.filters["variable_name"] = formatter.variable_name
-    env.filters["is_primitive"] = utils.is_primitive
+    env.filters["parameter_schema"] = openapi.parameter_schema
+    env.filters["parameters"] = openapi.parameters
     env.filters["is_json_api"] = openapi.is_json_api
-    env.filters["tf_sort_params_by_type"] = openapi.tf_sort_params_by_type
-    env.filters["tf_sort_properties_by_type"] = openapi.tf_sort_properties_by_type
+    env.filters["capitalize"] = utils.capitalize
+    env.filters["is_primitive"] = utils.is_primitive
+    env.filters["response_type"] = type.get_type_for_response
+    env.filters["return_type"] = type.return_type
+    env.filters["tf_sort_params_by_type"] = type.tf_sort_params_by_type
+    env.filters["tf_sort_properties_by_type"] = type.tf_sort_properties_by_type
 
     env.globals["enumerate"] = enumerate
     env.globals["get_name"] = openapi.get_name
-    env.globals["get_type_for_parameter"] = openapi.get_type_for_parameter
-    env.globals["get_type"] = openapi.type_to_go
-    env.globals["get_terraform_schema_type"] = formatter.get_terraform_schema_type
     env.globals["get_terraform_primary_id"] = openapi.get_terraform_primary_id
     env.globals["json_api_attributes_schema"] = openapi.json_api_attributes_schema
+    env.globals["get_terraform_schema_type"] = formatter.get_terraform_schema_type
+    env.globals["get_type_for_parameter"] = type.get_type_for_parameter
+    env.globals["get_type"] = type.type_to_go
 
     env.globals["GET_OPERATION"] = utils.GET_OPERATION
     env.globals["CREATE_OPERATION"] = utils.CREATE_OPERATION
